完善 还款记录数据推送

+++ src/main/java/com/utour/youdai/admin/project/fi/service/ILoanRepaymentActualService.java
@@ -63,5 +63,5 @@ public interface ILoanRepaymentActualService {
      */
     public int deleteLoanRepaymentActualById(Long id);
 
-    int save(JSONObject jsonObject);
+    Long save(JSONObject jsonObject);
 }
